main

+++ internal/components/receivers/single_endpoint_receiver_test.go
@@ -1,16 +1,5 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package receivers_test
 
@@ -64,7 +53,7 @@ func TestDownstreamParsers(t *testing.T) {
 		desc             string
 		receiverName     string
 		parserName       string
-		defaultPort      int
+		defaultPort      int32
 		listenAddrParser bool
 	}{
 		{"zipkin", "zipkin", "__zipkin", 9411, false},
@@ -118,7 +107,7 @@ func TestDownstreamParsers(t *testing.T) {
 				assert.NoError(t, err)
 				assert.Len(t, ports, 1)
 				assert.EqualValues(t, tt.defaultPort, ports[0].Port)
-				assert.Equal(t, naming.PortName(tt.receiverName, int32(tt.defaultPort)), ports[0].Name)
+				assert.Equal(t, naming.PortName(tt.receiverName, tt.defaultPort), ports[0].Name)
 			})
 
 			t.Run("allows port to be overridden", func(t *testing.T) {
@@ -142,7 +131,7 @@ func TestDownstreamParsers(t *testing.T) {
 				assert.NoError(t, err)
 				assert.Len(t, ports, 1)
 				assert.EqualValues(t, 65535, ports[0].Port)
-				assert.Equal(t, naming.PortName(tt.receiverName, int32(tt.defaultPort)), ports[0].Name)
+				assert.Equal(t, naming.PortName(tt.receiverName, tt.defaultPort), ports[0].Name)
 			})
 
 			t.Run("returns a default config", func(t *testing.T) {
